requirements.txt
Django==1.5.1
Mezzanine==1.4.16
MySQL-python==1.2.4
Pillow==2.2.1
South==0.8.2
argparse==1.2.1
bleach==1.2.2
boto==2.13.3
distribute==0.7.3
django-dbbackup==1.80.2
django-embed-video==0.6
django-extensions==1.2.2
django-registration==1.0
django-registration-defaults==0.4.3
dropbox==1.6
filebrowser-safe==0.2.28
grappelli-safe==0.2.22
html5lib==0.95
mezzanine-wymeditor==0.2
oauthlib==0.6.0
pytz==2013.7
requests==1.2.3
requests-oauthlib==0.3.3
six==1.4.1
wsgiref==0.1.2
